maps: Use pointer receivers for MapValue types.
bpf.MapValue interface function GetValuePtr() returns a pointer to a
new temporary if the function receiver is a value rather than a
pointer.

lxcmap, ipcache, and lbmap were also using value receivers for their
implementations of MapValue interface. The problem with this is that
any lookups would fail to return the actual value, as the
bpf.LookupElement would write the value into a temporary unaccessible
to the caller. No such lookups were performed, so this did not cause
any problems in practice. Fix the implementations to prevent future
problems.

This fix is otherwise low risk, but it has happened earlier in
development that GetValuePtr() implmentations were not fixed properly
and a pointer to the pointer receiver was returned. This is not
noticed by the compiler, and would result in garbage data being
written to/read from the bpf maps.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>